feat: overhaul review system with unified UI and star ratings - 마이페이지 리뷰 목록 구현 및 0.5 단위 별점 시스템 적용

+++ backend/community/models.py
@@ -11,7 +11,7 @@ class Review(models.Model):
     
     title = models.CharField(max_length=100) # 제목
     content = models.TextField()             # 내용
-    rank = models.IntegerField(default=5)    # 평점 (아까 프론트에서 rank 보냈죠?)
+    rank = models.FloatField(default=5)    # 평점 (아까 프론트에서 rank 보냈죠?)
     
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
